+++ src/compute-engine/global-types.ts
@@ -34,6 +34,8 @@ export type {
   SimplifyOptions,
   ExplainOptions,
   SymbolDefinition,
+  SymbolDefinitionInput,
+  PartialSymbolDefinition,
   SymbolDefinitions,
   LibraryDefinition,
   AngularUnit,